Refactor quick actions and Timodoro layout
